Fix CI build: Add static_cast for idx in ValueBoundsConstraintSet
- Cast idx from size_t to unsigned for ValueBoundsConstraintSet::computeConstantBound
- This matches the expected type signature {Value, unsigned}

+++ compiler/src/iree/compiler/Codegen/Utils/Utils.cpp
@@ -1962,7 +1962,7 @@ std::optional<VectorizationTileSizes> inferSizesFromIR(Value val) {
           if (ShapedType::isDynamic(dim)) {
             FailureOr<int64_t> maybeDimBound =
                 ValueBoundsConstraintSet::computeConstantBound(
-                    presburger::BoundType::UB, {val, idx},
+                    presburger::BoundType::UB, {val, static_cast<unsigned>(idx)},
                     /*stopCondition=*/nullptr, /*closedUB=*/true);
             if (failed(maybeDimBound)) {
               LDBG() << "failed to infer bounds for dynamic dim";
